+++ include/api/APIHooks.h
@@ -1,0 +1,52 @@
+/** @file
+
+  Internal SDK stuff
+
+  @section license License
+
+  Licensed to the Apache Software Foundation (ASF) under one
+  or more contributor license agreements.  See the NOTICE file
+  distributed with this work for additional information
+  regarding copyright ownership.  The ASF licenses this file
+  to you under the Apache License, Version 2.0 (the
+  "License"); you may not use this file except in compliance
+  with the License.  You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+ */
+
+#pragma once
+
+#include "APIHook.h"
+
+#include "ts/InkAPIPrivateIOCore.h"
+
+#include "tscore/List.h"
+
+/// A collection of API hooks.
+class APIHooks
+{
+public:
+  void append(INKContInternal *cont);
+  /// Get the first hook.
+  APIHook *head() const;
+  /// Remove all hooks.
+  void clear();
+  /// Check if there are no hooks.
+  bool is_empty() const;
+
+private:
+  Que(APIHook, m_link) m_hooks;
+};
+
+inline bool
+APIHooks::is_empty() const
+{
+  return nullptr == m_hooks.head;
+}
