+++ include/api/InkAPIInternal.h
@@ -33,6 +33,10 @@
 #include "I_Tasks.h"
 #include "Plugin.h"
 
+#include "api/APIHook.h"
+#include "api/APIHooks.h"
+#include "api/FeatureAPIHooks.h"
+
 #include "ts/InkAPIPrivateIOCore.h"
 #include "ts/experimental.h"
 
@@ -102,205 +106,6 @@ struct HttpAltInfo {
   float m_qvalue;
 };
 
-enum APIHookScope {
-  API_HOOK_SCOPE_NONE,
-  API_HOOK_SCOPE_GLOBAL,
-  API_HOOK_SCOPE_LOCAL,
-};
-
-/// A single API hook that can be invoked.
-class APIHook
-{
-public:
-  INKContInternal *m_cont;
-  int invoke(int event, void *edata) const;
-  APIHook *next() const;
-  APIHook *prev() const;
-  LINK(APIHook, m_link);
-
-  // This is like invoke(), but allows for blocking on continuation mutexes.  It is a hack, calling it can block
-  // the calling thread.  Hooks that require this should be reimplemented, modeled on the hook handling in HttpSM.cc .
-  // That is, try to lock the mutex, and reschedule the continuation if the mutex cannot be locked.
-  //
-  int blocking_invoke(int event, void *edata) const;
-};
-
-/// A collection of API hooks.
-class APIHooks
-{
-public:
-  void append(INKContInternal *cont);
-  /// Get the first hook.
-  APIHook *head() const;
-  /// Remove all hooks.
-  void clear();
-  /// Check if there are no hooks.
-  bool is_empty() const;
-
-private:
-  Que(APIHook, m_link) m_hooks;
-};
-
-inline bool
-APIHooks::is_empty() const
-{
-  return nullptr == m_hooks.head;
-}
-
-/** Container for API hooks for a specific feature.
-
-    This is an array of hook lists, each identified by a numeric identifier (id). Each array element is a list of all
-    hooks for that ID. Adding a hook means adding to the list in the corresponding array element. There is no provision
-    for removing a hook.
-
-    @note The minimum value for a hook ID is zero. Therefore the template parameter @a N_ID should be one more than the
-    maximum hook ID so the valid ids are 0..(N-1) in the standard C array style.
- */
-template <typename ID, ///< Type of hook ID
-          int N        ///< Number of hooks
-          >
-class FeatureAPIHooks
-{
-public:
-  FeatureAPIHooks();  ///< Constructor (empty container).
-  ~FeatureAPIHooks(); ///< Destructor.
-
-  /// Remove all hooks.
-  void clear();
-  /// Add the hook @a cont to the end of the hooks for @a id.
-  void append(ID id, INKContInternal *cont);
-  /// Get the list of hooks for @a id.
-  APIHook *get(ID id) const;
-  /// @return @c true if @a id is a valid id, @c false otherwise.
-  static bool is_valid(ID id);
-
-  /// Invoke the callbacks for the hook @a id.
-  void invoke(ID id, int event, void *data);
-
-  /// Fast check for any hooks in this container.
-  ///
-  /// @return @c true if any list has at least one hook, @c false if
-  /// all lists have no hooks.
-  bool has_hooks() const;
-
-  /// Check for existence of hooks of a specific @a id.
-  /// @return @c true if any hooks of type @a id are present.
-  bool has_hooks_for(ID id) const;
-
-  /// Get a pointer to the set of hooks for a specific hook @id
-  APIHooks const *operator[](ID id) const;
-
-private:
-  bool m_hooks_p = false; ///< Flag for (not) empty container.
-  /// The array of hooks lists.
-  APIHooks m_hooks[N];
-};
-
-template <typename ID, int N> FeatureAPIHooks<ID, N>::FeatureAPIHooks() {}
-
-template <typename ID, int N> FeatureAPIHooks<ID, N>::~FeatureAPIHooks()
-{
-  this->clear();
-}
-
-// The APIHooks::clear() method can't be inlined (easily), and we end up calling
-// clear() very frequently (it's used in a number of features). A rough estimate
-// is that we may call APIHooks::clear() as much as 230x per transaction (there's
-// 180 additional APIHooks that should be eliminated in a different PR). This
-// code at least avoids calling this function for a majority of the cases.
-// Before this code, APIHooks::clear() would show up as top 5 in perf top.
-template <typename ID, int N>
-void
-FeatureAPIHooks<ID, N>::clear()
-{
-  if (m_hooks_p) {
-    for (auto &h : m_hooks) {
-      if (!h.is_empty()) {
-        h.clear();
-      }
-    }
-    m_hooks_p = false;
-  }
-}
-
-template <typename ID, int N>
-void
-FeatureAPIHooks<ID, N>::append(ID id, INKContInternal *cont)
-{
-  if (is_valid(id)) {
-    m_hooks_p = true;
-    m_hooks[id].append(cont);
-  }
-}
-
-template <typename ID, int N>
-APIHook *
-FeatureAPIHooks<ID, N>::get(ID id) const
-{
-  return likely(is_valid(id)) ? m_hooks[id].head() : nullptr;
-}
-
-template <typename ID, int N>
-APIHooks const *
-FeatureAPIHooks<ID, N>::operator[](ID id) const
-{
-  return likely(is_valid(id)) ? &(m_hooks[id]) : nullptr;
-}
-
-template <typename ID, int N>
-void
-FeatureAPIHooks<ID, N>::invoke(ID id, int event, void *data)
-{
-  if (likely(is_valid(id))) {
-    m_hooks[id].invoke(event, data);
-  }
-}
-
-template <typename ID, int N>
-bool
-FeatureAPIHooks<ID, N>::has_hooks() const
-{
-  return m_hooks_p;
-}
-
-template <typename ID, int N>
-bool
-FeatureAPIHooks<ID, N>::is_valid(ID id)
-{
-  return 0 <= id && id < N;
-}
-
-class HttpAPIHooks : public FeatureAPIHooks<TSHttpHookID, TS_HTTP_LAST_HOOK>
-{
-};
-
-class TSSslHookInternalID
-{
-public:
-  explicit constexpr TSSslHookInternalID(TSHttpHookID id) : _id(id - TS_SSL_FIRST_HOOK) {}
-
-  constexpr
-  operator int() const
-  {
-    return _id;
-  }
-
-  static const int NUM = TS_SSL_LAST_HOOK - TS_SSL_FIRST_HOOK + 1;
-
-  constexpr bool
-  is_in_bounds() const
-  {
-    return (_id >= 0) && (_id < NUM);
-  }
-
-private:
-  const int _id;
-};
-
-class SslAPIHooks : public FeatureAPIHooks<TSSslHookInternalID, TSSslHookInternalID::NUM>
-{
-};
-
 class LifecycleAPIHooks : public FeatureAPIHooks<TSLifecycleHookID, TS_LIFECYCLE_LAST_HOOK>
 {
 };
@@ -349,6 +154,8 @@ public:
 private:
   std::unordered_map<std::string, INKContInternal *> cb_table;
 };
+
+#include "HttpAPIHooks.h"
 
 class HttpHookState
 {
@@ -403,7 +210,5 @@ HttpHookState::id() const
 
 void api_init();
 
-extern HttpAPIHooks *http_global_hooks;
 extern LifecycleAPIHooks *lifecycle_hooks;
-extern SslAPIHooks *ssl_hooks;
 extern ConfigUpdateCbTable *global_config_cbs;
