+++ iocore/eventsystem/CMakeLists.txt
@@ -50,6 +50,11 @@ target_link_libraries(inkevent
         yaml-cpp::yaml-cpp # transitive
 )
 
+set_target_properties(inkevent
+    PROPERTIES
+        POSITION_INDEPENDENT_CODE TRUE
+)
+
 if(TS_USE_HWLOC)
         target_link_libraries(inkevent PRIVATE hwloc::hwloc)
 endif()
