+++ iocore/net/Net.cc
@@ -28,6 +28,7 @@
 
 ************************************************************************/
 
+#include "SSLAPIHooks.h"
 #include "P_Net.h"
 #include <utility>
 
@@ -120,6 +121,7 @@ ink_net_init(ts::ModuleVersion version)
   if (!init_called) {
     configure_net();
     register_net_stats();
+    init_global_ssl_hooks();
   }
 
   init_called = 1;
