+++ iocore/net/P_Connection.h
@@ -49,6 +49,8 @@
 
 #pragma once
 
+#include "I_NetProcessor.h"
+
 #include "tscore/ink_platform.h"
 
 struct NetVCOptions;
