+++ iocore/net/unit_tests/unit_test_main.cc
@@ -26,6 +26,7 @@
 #include "I_Thread.h"
 #include "P_SSLConfig.h"
 #include "records/I_RecordsConfig.h"
+#include "SSLAPIHooks.h"
 #include "tscore/BaseLogFile.h"
 #include "tscore/Diags.h"
 #include "tscore/I_Layout.h"
@@ -56,6 +57,7 @@ public:
     main_thread->set_specific();
 
     SSLConfig::startup();
+    init_global_ssl_hooks();
   }
 
   void
