+++ proxy/CMakeLists.txt
@@ -21,7 +21,8 @@ add_library(proxy STATIC
         CacheControl.cc
         ControlBase.cc
         ControlMatcher.cc
-	    HostStatus.cc
+        HostStatus.cc
+        HttpAPIHooks.cc
         IPAllow.cc
         ParentConsistentHash.cc
         ParentRoundRobin.cc
@@ -64,6 +65,7 @@ target_link_libraries(proxy
     PUBLIC
         ts::inkcache
         ts::inkevent
+        ts::tsapicore
         ts::tscore
     PRIVATE
         ts::inkutils
