+++ proxy/http/CMakeLists.txt
@@ -61,6 +61,7 @@ target_link_libraries(http
         ts::inkevent
         ts::inkhostdb
         ts::proxy
+        ts::tsapicore
         ts::tscore
     PRIVATE
         ts::inkcache
