+++ proxy/http/remap/CMakeLists.txt
@@ -36,7 +36,10 @@ add_library(http_remap STATIC
 )
 add_library(ts::http_remap ALIAS http_remap)
 
-target_include_directories(http_remap PRIVATE
+target_include_directories(http_remap
+    PUBLIC
+        ${CMAKE_CURRENT_SOURCE_DIR}
+    PRIVATE
         ${IOCORE_INCLUDE_DIRS}
         ${PROXY_INCLUDE_DIRS}
         ${CMAKE_SOURCE_DIR}/src/records
