+++ proxy/http/remap/unit-tests/nexthop_test_stubs.cc
@@ -60,10 +60,6 @@ HttpCacheAction::cancel(Continuation *c)
 {
 }
 PostDataBuffers::~PostDataBuffers() {}
-void
-APIHooks::clear()
-{
-}
 
 HttpTunnel::HttpTunnel() {}
 HttpCacheSM::HttpCacheSM() {}
